[No issue] Rename deck import model files
